placed all files under their respective folders
