+++ c++/Arranging Coins Implementation.cpp
