+++ c++/Rat in a Maze.cpp
