+++ c++/gcd.cpp
