+++ python/linkedlist.py
